requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytest==8.0.0
flask-cors==4.0.0
pyngrok==7.1.5